requirements.txt
# Core web framework
Flask==2.3.3
Werkzeug==2.3.7

# HTTP requests and utilities
requests==2.31.0
python-dotenv==1.0.0

# Image processing
Pillow==10.0.1
pillow-heif==0.13.0

# Face recognition (optional - will fallback if not available)
face_recognition==1.3.0
opencv-python==4.8.1.78

# RAW image processing (optional)
rawpy==0.18.1

# Database and storage
firebase-admin
firebase==1.2.0

# Scientific computing
numpy==1.24.3

# Concurrent processing (built-in but explicit for clarity)
# concurrent.futures is part of Python 3.2+ standard library

# Development and deployment
gunicorn==21.2.0

# Optional: For better performance
psutil==5.9.6

# Advanced face recognition improvements (optional - will fallback if not available)
facenet-pytorch==2.5.3
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
albumentations==1.3.1

# KWIKPIC KILLER DEPENDENCIES - Beat 99.9% accuracy!
retinaface==0.0.13
mediapipe==0.10.8
ultralytics==8.0.196
opencv-contrib-python==4.8.1.78

# InsightFace (TensorRT/ONNX-based RetinaFace alternative)
insightface==0.7.3
# CPU runtime (use onnxruntime-gpu if you have NVIDIA GPU drivers/CUDA)
onnxruntime==1.18.1